move superuser users table to livecomponent
