+++ lib/lightning_web/live/user_live/index.html.heex
@@ -13,14 +13,11 @@
     </LayoutComponents.header>
   </:header>
   <LayoutComponents.centered>
-    <.users_table
-      socket={@socket}
-      live_action={@live_action}
+    <.live_component
+      id="users-table-live-component"
+      module={LightningWeb.UserLive.TableComponent}
       delete_user={assigns[:delete_user]}
-      users={@users}
-      sort_key={@sort_key}
-      sort_direction={@sort_direction}
-      filter={@filter}
+      live_action={@live_action}
     />
   </LayoutComponents.centered>
 </LayoutComponents.page_content>
